test(playground): keep inspector files under the source-length budget
